Respect reuse-db for clickhouse

+++ ee/conftest.py
@@ -13,7 +13,7 @@ from posthog.settings import (
 
 
 @pytest.fixture(scope="package")
-def django_db_setup(django_db_setup):
+def django_db_setup(django_db_setup, django_db_keepdb):
     database = Database(
         CLICKHOUSE_DATABASE,
         db_url=CLICKHOUSE_HTTP_URL,
@@ -21,18 +21,24 @@ def django_db_setup(django_db_setup):
         password=CLICKHOUSE_PASSWORD,
         verify_ssl_cert=CLICKHOUSE_VERIFY,
     )
-    try:
-        database.drop_database()
-    except:
-        pass
-    database.create_database()
-    database.migrate("ee.clickhouse.migrations")
+
+    if not django_db_keepdb:
+        try:
+            database.drop_database()
+        except:
+            pass
+
+    if not django_db_keepdb or not database.db_exists:
+        database.create_database()
+        database.migrate("ee.clickhouse.migrations")
+
     # Make DELETE / UPDATE synchronous to avoid flaky tests
     sync_execute("SET mutations_sync = 1")
 
     yield
 
-    try:
-        database.drop_database()
-    except:
-        pass
+    if not django_db_keepdb:
+        try:
+            database.drop_database()
+        except:
+            pass
